+++ CHANGELOG.md
@@ -8,6 +8,7 @@ This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html
 ### Major Changes
 - Added support to ID Pools.
 - Added New Resource Appliance Configuration Timeconfig
+- Added New Resource Appliance SSH Access
 - Added support to Connection Template
 
 ### Bug fixes & Enhancements: